requirements.txt
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-anthropic>=0.3.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
python-dotenv>=1.0.0
python-pptx>=1.0.0
sentence-transformers>=3.0.0
flask>=3.0.0
